create mate dashboard ui
