+++ src/main/kotlin/ui/views/task/TaskUIController.kt
@@ -5,5 +5,4 @@ import java.util.*
 interface TaskUIController {
     fun editTask(taskId: UUID)
     fun deleteTask(taskId: UUID)
-    fun viewTaskDetails(taskId: UUID)
 }
